refactor: 重构edgeone构建脚本
