+++ cloud-functions/mock/bili-bgm.js
