+++ cloud-functions/mock/index.js
