--- node-functions/shared/index.js
@@ -1,1 +1,0 @@
-export * from './utils.js'
