+++ packages/api/src/edgeone.ts
@@ -1,69 +1,80 @@
 import path from 'node:path'
-import { existsSync, mkdirSync, readFile, rmSync, writeFile } from 'node:fs'
+import { execSync } from 'node:child_process'
+import { existsSync, mkdirSync, rmSync } from 'node:fs'
+import { readFile, writeFile } from 'node:fs/promises'
 import { fileURLToPath } from 'node:url'
-import { promisify } from 'node:util'
-
-const readFileAsync = promisify(readFile)
-const writeFileAsync = promisify(writeFile)
+import { createRequire } from 'node:module'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
 const projectRootDir = path.resolve(__dirname, '../../../')
-const outputDir = path.join(projectRootDir, 'node-functions')
+const outputDir = path.join(projectRootDir, 'cloud-functions')
 const srcDir = path.resolve(__dirname, '..', 'src')
-const customSrcDir = path.resolve(projectRootDir, 'packages/bilibili-bangumi-component/src/shared')
 
-function transform(code: string) {
-  return code
-    .replace(/import .* from '[^']+';/g, '')
-    .replace(/import \{[^}]+\} from '[^']+'/g, '')
-    .replace(/import \* as \w+ from '[^']+'/g, '')
-    .replace(/import type \{[^}]*\} from '[^']+'/g, '')
-    .replace(/export async function GET\(request: Request\)/g, 'export default function onRequestGet(context)')
-    .replace(/process\.env/g, 'context.env')
-    .replace(/\?:/g, ':')
-    .replace(/ as any/g, '')
-    .replace(/const isMock = true/g, 'const isMock = true')
-    .replace(/const isMock = false/g, 'const isMock = false')
-    .replace(/export function generateRes\(params\)/g, 'export function generateRes(params) {\n  return Response.json(params, { status: params.code })')
-    .replace(/\): ResponseData/g, ')')
-    .replace(/: \{ pageNumber: number, pageSize: number \}/g, '')
-    .replace(/\?.data\[\]/g, '?.data')
-    .replace(/\?.list\[\]/g, '?.list')
-    .replace(/ as (?:any|string)(?:\[\])?/g, '')
-    .replace(/: any([,)])/g, '$1')
-    .replace(/<ListItem>/g, '')
-    .replace(/: LabelItem\[\]|: BilibiliQuery|: BgmQuery|: NodeJS\.ProcessEnv|: Record<[^>]+>/g, '')
-    .replace(/export interface CustomData[\s\S]*?^\}/gm, '')
-    .replace(/export interface CollectionData[\s\S]*?^\}/gm, '')
-    .replace(/\/\*\*[\s\S]*?\*\//g, '')
-    .replace(/export function (customHandler|generateEmpty)/g, 'function $1')
-    .replace(/: CustomQuery|: CustomData|: Response|: number|: ListItem\[\]/g, '')
-    .replace(/^['\.\/].*/gm, '')
-    .replace(/^$/gm, '')
-    .replace(/\n{2,}/g, '\n')
-    .trim()
+const sharedUtilsPath = path.resolve(projectRootDir, 'packages/bilibili-bangumi-component/src/shared/utils').replace(/\\/g, '/')
+const bilibiliPath = path.resolve(projectRootDir, 'packages/api/src/bilibili').replace(/\\/g, '/')
+const bgmPath = path.resolve(projectRootDir, 'packages/api/src/bgm').replace(/\\/g, '/')
+const apiUtilsPath = path.resolve(projectRootDir, 'packages/api/src/shared/utils').replace(/\\/g, '/')
+const apiSharedPath = path.resolve(projectRootDir, 'packages/api/src/shared').replace(/\\/g, '/')
+
+function getEntryCode() {
+  return `
+import { parseSearchParams, generateRes } from '${sharedUtilsPath}'
+import { handler as bilibili } from '${bilibiliPath}'
+import { handler as bgm } from '${bgmPath}'
+import { handleQuery } from '${apiUtilsPath}'
+import { customHandler } from '${apiSharedPath}'
+
+function getRouteType(pathname) {
+  const segments = pathname.split('/').filter(Boolean)
+  const lastSegment = segments[segments.length - 1]
+  return ['bilibili', 'bgm', 'custom'].includes(lastSegment) ? lastSegment : null
 }
 
-function wrapApi(code: string, imports: string, name?: string, isEnv = true) {
-  return `${imports}
-
-${code}
-
-export default function onRequestGet(context) {
-  const { request } = context${isEnv
-? `
-  const env = context.env || {}`
-: ''}
+export default async function onRequestGet(context) {
+  const { request } = context
+  const env = context.env || {}
   const url = new URL(request.url)
   const query = handleQuery(parseSearchParams(url))
+  const routeType = getRouteType(url.pathname)
 
-  ${name ? `if (env?.MOCK === 'true')\n    return Response.json(mockDataMap.${name})\n\n  return handler(query, env)` : `return customHandler(query, customData)`}
-}`
+  let mockDataMap = { bilibili: {}, bgm: {} }
+  let customData = {}
+  try {
+    const mockModule = await import('../mock/index.js')
+    mockDataMap = mockModule.mockDataMap || mockDataMap
+    customData = mockModule.customData || customData
+  } catch {}
+
+  if (!routeType) {
+    return generateRes({
+      code: 404,
+      message: 'Not found. Available routes: /api/bilibili, /api/bgm, /api/custom',
+      data: {},
+    })
+  }
+
+  if (env?.MOCK === 'true' && routeType !== 'custom')
+    return Response.json(mockDataMap[routeType])
+
+  switch (routeType) {
+    case 'bilibili':
+      return await bilibili(query, env)
+    case 'bgm':
+      return await bgm(query, env)
+    case 'custom':
+      return customHandler(query, customData)
+    default:
+      return generateRes({ code: 404, message: 'Not found', data: {} })
+  }
+}
+`
 }
 
-function wrapMock(code: string, exports: string) {
-  return `${code.replace(/\/\* eslint-disable.*?\*\/\n?/g, '').trim()}\n\nexport { ${exports} }\n`
+function cleanDir(dir: string) {
+  if (existsSync(dir))
+    rmSync(dir, { recursive: true })
+  mkdirSync(dir, { recursive: true })
 }
 
 async function buildEdgeOne() {
@@ -72,40 +83,54 @@ async function buildEdgeOne() {
 
   const apiDir = path.resolve(outputDir, 'api')
   const mockDir = path.resolve(outputDir, 'mock')
+  const tempEntry = path.resolve(outputDir, '__edgeone_entry__.ts')
 
-  if (existsSync(apiDir))
-    rmSync(apiDir, { recursive: true })
-  if (existsSync(mockDir))
-    rmSync(mockDir, { recursive: true })
+  cleanDir(apiDir)
+  cleanDir(mockDir)
 
-  mkdirSync(apiDir, { recursive: true })
-  mkdirSync(mockDir, { recursive: true })
-
-  const [bilibiliCode, bgmCode, customCodeSrc, mockBiliBgmCode, mockCustomCode] = await Promise.all([
-    readFileAsync(path.resolve(srcDir, 'bilibili.ts'), 'utf-8'),
-    readFileAsync(path.resolve(srcDir, 'bgm.ts'), 'utf-8'),
-    readFileAsync(path.resolve(customSrcDir, 'custom.ts'), 'utf-8'),
-    readFileAsync(path.resolve(srcDir, 'mock/bili-bgm.js'), 'utf-8'),
-    readFileAsync(path.resolve(srcDir, 'mock/custom.js'), 'utf-8'),
+  const [mockBiliBgmCode, mockCustomCode] = await Promise.all([
+    readFile(path.resolve(srcDir, 'mock/bili-bgm.js'), 'utf-8'),
+    readFile(path.resolve(srcDir, 'mock/custom.js'), 'utf-8'),
   ])
 
-  const imports = {
-    bilibili: `import { mockDataMap } from '../mock/index.js'\nimport { generateRes, handleQuery, numberToZh, parseSearchParams, serializeSearchParams } from '../shared/index.js'`,
-    bgm: `import { mockDataMap } from '../mock/index.js'\nimport { generateRes, handleQuery, parseSearchParams, serializeSearchParams } from '../shared/index.js'`,
-    custom: `import { customData } from '../mock/index.js'\nimport { generateRes, handleQuery, parseSearchParams } from '../shared/index.js'`,
+  const require = createRequire(import.meta.url)
+  const esbuild = require('esbuild')
+
+  await writeFile(tempEntry, getEntryCode())
+
+  try {
+    await esbuild.build({
+      entryPoints: [tempEntry],
+      bundle: true,
+      format: 'esm',
+      platform: 'node',
+      target: 'node20',
+      outfile: path.resolve(apiDir, '[[default]].js'),
+      minify: false,
+      treeShaking: true,
+      charset: 'utf8',
+      external: ['./mock/index.js', '../mock/index.js', '../../mock/index.js'],
+      resolveExtensions: ['.ts', '.js', '.json'],
+      sourcemap: false,
+    })
+  }
+  finally {
+    if (existsSync(tempEntry))
+      rmSync(tempEntry)
   }
 
   await Promise.all([
-    writeFileAsync(path.resolve(apiDir, 'bilibili.js'), `${wrapApi(transform(bilibiliCode), imports.bilibili, 'bilibili', true)}\n`),
-    writeFileAsync(path.resolve(apiDir, 'bgm.js'), `${wrapApi(transform(bgmCode), imports.bgm, 'bgm', true)}\n`),
-    writeFileAsync(path.resolve(apiDir, 'custom.js'), `${wrapApi(transform(customCodeSrc), imports.custom, '', false)}\n`),
-    writeFileAsync(path.resolve(mockDir, 'bili-bgm.js'), wrapMock(mockBiliBgmCode, 'mockDataMap')),
-    writeFileAsync(path.resolve(mockDir, 'custom.js'), wrapMock(mockCustomCode, 'customData')),
-    writeFileAsync(path.resolve(mockDir, 'index.js'), `export { mockDataMap } from './bili-bgm.js'\nexport { customData } from './custom.js'\n`),
+    writeFile(path.resolve(mockDir, 'bili-bgm.js'), `${mockBiliBgmCode}\nexport { mockDataMap }\n`),
+    writeFile(path.resolve(mockDir, 'custom.js'), `${mockCustomCode}\nexport { customData }\n`),
+    writeFile(path.resolve(mockDir, 'index.js'), `export { mockDataMap } from './bili-bgm.js'\nexport { customData } from './custom.js'\n`),
   ])
 
   // eslint-disable-next-line no-console
-  console.info('EdgeOne build complete! Output:', path.resolve(outputDir))
+  console.info('EdgeOne build complete! Output:', outputDir)
+
+  execSync('npx eslint --fix cloud-functions/', { cwd: projectRootDir, stdio: 'pipe' })
+  // eslint-disable-next-line no-console
+  console.info('ESLint fix applied')
 }
 
 buildEdgeOne()
